+++ g3w-toolbox.js
@@ -32,18 +32,16 @@ import { checkSessionItems }                            from './utils/checkSessi
 import { unlinkRelation }                               from './utils/unlinkRelation';
 import { splitFeatures }                                from './utils/splitFeatures';
 import { isSameBaseGeometryType }                       from './utils/isSameBaseGeometryType';
-import { PickFeaturesInteraction }                      from './interactions/pickfeatures';
+import { PickFeaturesInteraction }                      from './actions/pick-features';
 
-import {
-  OpenFormStep,
-  SelectElementsStep,
-  PickFeatureStep,
-  AddFeatureStep,
-  MoveFeatureStep,
-  RotateFeatureStep,
-  ModifyGeometryVertexStep,
-  OpenTableStep,
-}                                                       from './workflows';
+import { OpenFormStep }                                 from './actions/open-form';
+import { SelectElementsStep }                           from './actions/select-elements';
+import { PickFeatureStep }                              from './actions/pick-feature';
+import { AddFeatureStep }                               from './actions/add-feature';
+import { MoveFeatureStep }                              from './actions/move-feature';
+import { RotateFeatureStep }                            from './actions/rotate-feature';
+import { ModifyGeometryVertexStep }                     from './actions/modify-geometry-vertex';
+import { OpenTableStep }                                from './actions/open-table';
 
 Object
   .entries({
